// x-pack/platform/plugins/shared/maps/public/reducers/map/types.d.ts
import type { KibanaExecutionContext } from '@kbn/core/public';
import type { Query } from '@kbn/data-plugin/common';
import type { Filter, ProjectRouting } from '@kbn/es-query';
import type { TimeRange } from '@kbn/es-query';
import type { DrawState, EditState, Goto, LayerDescriptor, MapCenter, MapExtent, MapSettings, Timeslice, TooltipState } from '../../../common/descriptor_types';
export interface MapExtentState {
    zoom: number;
    extent: MapExtent;
    center: MapCenter;
}
export type MapViewContext = MapExtentState & {
    buffer: MapExtent;
};
export type MapContext = Partial<MapViewContext> & {
    mouseCoordinates?: {
        lat: number;
        lon: number;
    };
    timeFilters?: TimeRange;
    timeslice?: Timeslice;
    query?: Query;
    filters: Filter[];
    embeddableSearchContext?: {
        query?: Query;
        filters: Filter[];
    };
    drawState?: DrawState;
    editState?: EditState;
    searchSessionId?: string;
    searchSessionMapBuffer?: MapExtent;
    projectRouting?: ProjectRouting;
};
export type MapState = {
    executionContext: KibanaExecutionContext;
    ready: boolean;
    mapInitError?: string | null;
    goto?: Goto | null;
    openTooltips: TooltipState[];
    mapState: MapContext;
    selectedLayerId: string | null;
    layerList: LayerDescriptor[];
    waitingForMapReadyLayerList: LayerDescriptor[];
    settings: MapSettings;
    __rollbackSettings: MapSettings | null;
    __pauseSyncData: boolean;
};